module-user_interface/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
asyncio==3.4.3
attrs==24.2.0
#av==14.0.1
bidict==0.23.1
blinker==1.9.0
click==8.1.7
discord.py==2.4.0
dnspython==2.7.0
eventlet==0.38.0
Flask==3.1.0
Flask-SocketIO==5.4.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
libarchive-c==5.1
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.2.0
opencv-python==4.10.0.84
picamera2==0.3.23
pidng==4.0.9
piexif==1.1.3
pillow==11.0.0
ply==3.11
propcache==0.2.1
python-dotenv==1.0.1
python-engineio==4.10.1
python-prctl==1.8.1
python-socketio==5.11.4
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
rpi-kms==0.1a1
rpi-libcamera==0.1a7
RPi.GPIO==0.7.1
setuptools==75.6.0
simple-websocket==1.1.0
simplejpeg==1.7.6
spidev==3.6
tqdm==4.67.1
v4l2-python3==0.3.5
#VimbaPython @ file:///home/raspberry/iibm/dev-webapp/VimbaPython-1.3.0-py3-none-any.whl#sha256=8ed623cb14fc03bd5082efd18e9098b7756cdbb54ada3063690e0ebf0abcb28b
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
yarl==1.18.3
